project completion

+++ src/main/webapp/managePhones.jsp
@@ -3,7 +3,6 @@
 <%@ page import="java.sql.ResultSet" %>
 <%@ page import="com.swapSmart.utils.DBConnection" %>
 <%@ include file="./components/adminNavbar.jsp" %>
-<%@ include file="./components/adminSidebar.jsp" %>
 <link rel="stylesheet" href="./assets/css/adminStyles.css">
 
 <h2 class="dashboard-heading">Manage Phones for Sale</h2>
@@ -21,13 +20,17 @@
     </thead>
     <tbody>
         <%
-            // Connection to the database
-            Connection conn = DBConnection.getConnection();
-            String sql = "SELECT * FROM phones_for_sale";
-            PreparedStatement stmt = conn.prepareStatement(sql);
-            ResultSet rs = stmt.executeQuery();
+            Connection conn = null;
+            PreparedStatement stmt = null;
+            ResultSet rs = null;
 
-            while (rs.next()) {
+            try {
+                conn = DBConnection.getConnection();
+                String sql = "SELECT * FROM phones_for_sale";
+                stmt = conn.prepareStatement(sql);
+                rs = stmt.executeQuery();
+
+                while (rs.next()) {
         %>
         <tr>
             <td><%= rs.getInt("id") %></td>
@@ -36,10 +39,18 @@
             <td><%= rs.getDouble("price") %></td>
             <td><%= rs.getBoolean("is_available") ? "Yes" : "No" %></td>
             <td>
-                <a href="deletePhone?id=<%= rs.getInt("id") %>" class="custom-btn danger">Delete</a>
+                <a href="editPhone.jsp?id=<%= rs.getInt("id") %>" class="custom-btn warning">Edit</a>
+                <a href="deletePhone?id=<%= rs.getInt("id") %>" class="custom-btn danger" onclick="return confirm('Are you sure you want to delete this phone?');">Delete</a>
             </td>
         </tr>
-        <% 
+        <%
+                }
+            } catch (Exception e) {
+                out.println("<tr><td colspan='6'>Error: " + e.getMessage() + "</td></tr>");
+            } finally {
+                if (rs != null) rs.close();
+                if (stmt != null) stmt.close();
+                if (conn != null) conn.close();
             }
         %>
     </tbody>
